requirements.txt
aiohttp==3.8.4
aiomysql==0.0.22
aiosignal==1.2.0
aiosmtplib==1.1.7
arrow==1.2.3
asgiref==3.4.1
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==22.2.0
beautifulsoup4==4.11.1
boto3==1.23.10
botocore==1.26.10
cachetools==4.2.4
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==3.0.1
click==7.1.2
colorama==0.4.5
crypto==1.4.1
cryptography==38.0.1
cryptor==0.0.1
cssselect==1.1.0
cssutils==2.3.1
curlify==2.2.1
DBUtils==1.3
defusedxml==0.7.1
et-xmlfile==1.1.0
facebook-business==16.0.2
Flask==1.1.4
Flask-Cors==3.0.10
frozenlist==1.2.0
future==0.18.3
google==3.0.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
idna==3.4
idna-ssl==1.1.0
importlib-metadata==4.8.3
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.0.1
multidict==5.2.0
mysqlclient==1.4.6
Naked==0.1.32
numpy==1.19.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
pandas==1.1.5
Pillow==6.2.2
premailer==3.10.0
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.15.0
pyOpenSSL==23.2.0
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==10.0.0
pytz==2022.2.1
PyYAML==6.0
qrcode==6.1
redis==4.3.5
requests==2.6.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
tablib==0.14.0
telegram==0.0.1
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.16
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.15.293
yarl==1.7.2
zipp==3.6.0
pymongo==4.6.1